Updated Intel Edison SensorTag sample to use connectthedots npm package
