--- Devices/DirectlyConnectedDevices/NodeJS/IntelEdisonSensorTag/ConnectTheDotsIntelEdisonSensorTag.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9092AA53-FB77-4645-B42D-1CCCA6BD08BD}") = "ConnectTheDotsIntelEdisonSensorTag", "ConnectTheDotsIntelEdisonSensorTag.njsproj", "{C300FE92-9367-4FF2-A309-306ADD5596F2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C300FE92-9367-4FF2-A309-306ADD5596F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C300FE92-9367-4FF2-A309-306ADD5596F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C300FE92-9367-4FF2-A309-306ADD5596F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C300FE92-9367-4FF2-A309-306ADD5596F2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
